feat: move code to lib
