+++ __tests__/questions.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable nada/path-case */
 const fs = require('fs');
-const questions = require('../questions.js');
+const questions = require('../lib/questions.js');
 
 describe('cz-customizable', () => {
   let config;
