+++ lib/build-commit.js
