+++ lib/logger.js
