+++ lib/read-config.js
