+++ standalone.js
@@ -4,7 +4,7 @@ const { execSync } = require('child_process');
 const inquirer = require('inquirer');
 
 const app = require('./index');
-const log = require('./logger');
+const log = require('./lib/logger');
 
 log.info('cz-customizable standalone version');
